# Add first pass at SparseArray class with examples