--- a/examples/py4dstem/cropped.ipynb
+++ b/examples/py4dstem/cropped.ipynb
@@ -56,19 +56,11 @@
     "    'data': data.data,\n",
     "    'scan_shape': data.scan_dimensions,\n",
     "    'frame_shape': data.frame_dimensions,\n",
+    "    'dtype': np.uint32,\n",
     "    'allow_full_expand': False,\n",
+    "    'sparse_slicing': True,\n",
     "}\n",
     "array = SparseArray(**kwargs)"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "# Full expansion of the array is not allowed, to ensure we don't ever accidentally do it\n",
-    "#array[:]"
    ]
   },
   {
@@ -158,8 +150,8 @@
    "outputs": [],
    "source": [
     "# Position a bright-field detector\n",
-    "x0,y0 = 300, 275\n",
-    "R = 90\n",
+    "x0,y0 = 302, 275\n",
+    "R = 40\n",
     "\n",
     "py4DSTEM.visualize.show_circ(sum_dp,0,2,center=(x0,y0),R=R,alpha=0.25)"
    ]
@@ -170,6 +162,12 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "# Need to turn off sparse slicing so py4dstem will get a dense array in this function\n",
+    "# FIXME: maybe we can override __mul__ and __rmul__ so multiplying a numpy array to\n",
+    "# a sparse array causes the sparse array to expand? That would remove the need to\n",
+    "# turn off sparse slicing here.\n",
+    "datacube.data.sparse_slicing = False\n",
+    "\n",
     "# Show the bright-field image\n",
     "BF_image = py4DSTEM.process.virtualimage.get_virtualimage_circ(datacube,x0,y0,R)\n",
     "py4DSTEM.visualize.show(BF_image,contrast='minmax')"