--- a/examples/py4dstem/cropped.ipynb
+++ b/examples/py4dstem/cropped.ipynb
@@ -38,7 +38,7 @@
     "import py4DSTEM\n",
     "\n",
     "from py4DSTEM.io import DataCube\n",
-    "from sparse_array import SparseArray\n",
+    "from stempy.io.sparse_array import SparseArray\n",
     "import stempy.io"
    ]
   },