--- a/examples/py4dstem/cropping_binning.ipynb
+++ b/examples/py4dstem/cropping_binning.ipynb
@@ -9,7 +9,7 @@
     "import numpy as np\n",
     "import matplotlib.pyplot as plt\n",
     "\n",
-    "from sparse_array import SparseArray\n",
+    "from stempy.io.sparse_array import SparseArray\n",
     "import stempy.io"
    ]
   },