--- a/examples/py4dstem/requirements.txt
+++ b/examples/py4dstem/requirements.txt
@@ -1,0 +1,1 @@
+py4DSTEM