… docs/source/stempy.io.rst
… examples/py4dstem/sparse_array.py