The SparseArray class can take electron counted data and use it to
perform operations such as min/max/mean/sum, indexing, and slicing.

Slicing and indexing operations are allowed, such as `array[3:60, 5, ::2, :3]`,
where the only the frames that get expanded are the ones that are needed.

Arithmetic operations can be performed, such as `np.sum(array, axis=(0, 1))`.
Note that only axis=(0, 1) is currently specialized for efficiency. If you
use an axis that isn't specialized, it is like calling
`np.sum(array[:], axis)`, where the array gets expanded to a full numpy array
and then the operation performed on it. We can specialize more axes options
as needed.

The class also accepts an argument that indicates whether or not full expansion
(such as `array[:]`) is allowed. If `False`, then an exception will be raised
if a full expansion is attempted. This is helpful to ensure that functions
that are called do not attempt a full expansion.

The class also accepts a `dtype`, which is the type that will be used when
an expansion happens. For things like a sum, it is important to pick a dtype
large enough so that it won't overflow.

The examples are taken from the py4dstem quick start and modified slightly to
use our SparseArray class. Some things to note:

1. We are using sum() instead of max() because max() for our example is mostly
   all ones with a few zeros (not very visually appealing).
2. The binning and cropping steps are removed. They work, but they do so in
   py4DSTEM by slicing and re-assigning (such as
   `datacube.data = datacube.data[:, :, 3:20, 4:15]`). Since slicing the sparse
   array returns a numpy array, this means the sparse array will get replaced
   by a numpy array, which we might not want for this example.
3. The saving to HDF5 at the end doesn't work because py4DSTEM doesn't
   recognize our object type.
4. We have to swap scan dimensions for the full battery example for the image
   to come out correctly. We might need to make sure we are using the dimensions
   correctly.

The two jupyter notebooks with their output are attached in this [examples.zip](https://github.com/OpenChemistry/stempy/files/5805251/examples.zip) file.

Let us know what you think, @ercius