@ercius 2021-01-13
This looks great.  Ill try it out soon.

> The class also accepts an argument that indicates whether or not full expansion
(such as array[:]) is allowed. If False, then an exception will be raised
if a full expansion is attempted. This is helpful to ensure that functions
that are called do not attempt a full expansion.

I was really worried about the possibility to nuke the users RAM, but this is a great way to avoid this. Thanks for putting this in there. `False` should be the default value.