@ercius 2021-01-30
@psavery Is there a docker image I can download to test this? The example notebooks look like they are just what we want to accomplish. The dataset has insufficient signal for the Bragg disc detection. I will have to get a different dataset to test for that or try out some different py4dstem functionality.