@cjh1 2021-02-01
@ercius Its really only a single python file ( `spare_array.py` ), it currently does need any of the other stempy machinery. So you can either get the file and notebook from the PR or we can send it to you?