@ercius 2021-02-01
I can grab it. Thanks!

On Mon, Feb 1, 2021, 5:51 AM Chris Harris <notifications@github.com> wrote:

> @ercius <https://github.com/ercius> Its really only a single python file
> ( spare_array.py ), it currently does need any of the other stempy
> machinery. So you can either get the file and notebook from the PR or we
> can send it to you?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/OpenChemistry/stempy/pull/191#issuecomment-770871411>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADXORFGUXKC2ZRWM3EI5JA3S42WWVANCNFSM4V75EVBQ>
> .
>