@ercius 2021-02-07
@psavery This works really well. Nice work! Should we discuss the next steps to improving this and what to add?