@cjh1 2021-02-08
@ercius Glad this works for you. Yes, we should discuss what other things to add and where the code should end up living ...