@ercius 2021-02-10
Those are good questions.

More features to add are:
1. binning the diffraction data to change the frame dimensions in place. 
2.  cropping the data in both diffraction and real space
I have python code for both of these which I can share.

As for where it should live: I would like it to be easily accessible and installable. We should talk about how to make stempy more accessible or maybe to start a different repo or add it to an existing repo (ncempy?). Im not sure of the right way to do this and would like to discuss that.