@cjh1 2021-02-11
@ercius Just so we are on the same page. For 1. and 2. these are operations supported in py4DSTEM? As in we would like to support them with our sparse array? Sharing the code would be great.

I agree that making stempy more accessible would be great. I know we have talked about pushing packages to pypi in the past but put it off because of the potential time sink. @psavery has setup several projects recently to build for pypi and I think we have streamlined the process. Given this it may be worth revisiting this for stempy. I think with a few hours investment we would know pretty soon if its going to be possible in a tractable amount of time. If this didn't work out the sparse array code could also end up in py4DSTEM itself.