@ercius 2021-02-11
For 1 and 2 you can use functions in py4dstem, but I think it would require working on the dense data. I was planning for 1 to add a function to bin the data in place. You can change the way the sparse --> dense operation happens in order to reduce the frame_dimensions. It would be similar to how numpy arrays can be reshaped by only changing the shape and not actually changing the underlying data. For 2 I think this would also be built in but probably would need to return a new array. Im not sure how I would want that to work.

Lets see if we can get it distributable in pypi. That is certainly worth some effort at this point.