@psavery 2021-02-16
Hi Peter @ercius,

I've added cropping (via slicing) and binning for frames. I've attached
[an example notebook with output](https://github.com/OpenChemistry/stempy/files/5991875/cropping_binning_with_output.ipynb.zip) and more details are below. Let us
know what you think.


Changes:
========

1. The SparseArray class now has an attribute called "sparse_slicing",
   which defaults to True. If True, then slicing a SparseArray will
   return a new SparseArray with the underlying data modified (except
   when the slicing results in a single frame, in which case it returns
   a dense array for the frame).

   This allows for cropping and striding the data via slicing. If
   "sparse_slicing" is False, then a dense array will be returned upon
   slicing.

   Since py4dstem also crops data via slicing, cropping a SparseArray
   with their functions should yield the same result.

2. Previously, the math functions and expansion functions assumed that
   the presense of a sparse value in a frame indicated that the pixel
   value is 1. However, this is not the case, because the frame can
   have duplicated sparse values, which indicates a higher pixel value.
   The underlying math and expansion functions have now been modified
   to reflect this, and they now count the number of sparse values in
   all cases. This is important also for binning, where duplicated
   values are needed to indicate greater intensity in a pixel.

   This results in a little bit of a slowdown for the functions, but
   accuracy is more important. Current timings are pasted below.

3. A bin_frames() function was added to bin the frames. This either
   returns a new SparseArray (default), or if "in_place" == True,
   it modifies the SparseArray in place and returns "self".

   We currently can't use the py4dstem functions to bin the data,
   because they perform a reshape and sum. We can only bin the frames
   via this "bin_frames()" function.


Questions:
==========

1. There is a dtype in the SparseArray class which indicates what
   dtype should be used for expansion and math functions. This
   currently defaults to np.uint8, which causes integer overflows
   for most of our examples, for things like summing. Should we
   increase the default to something else, like np.uint32?

2. Adding a "bin_scans()" function might not be too hard. Should we
   add one of those too, to bin the scans?


Current math timings:
=====================

Small dataset:
data.frame_dimensions=[576, 576]
data.scan_dimensions=[64, 64]

min:  3.90 s
max:  0.92 s
sum:  0.86 s
mean: 0.84 s


Large dataset:
data.frame_dimensions=[576, 576]
data.scan_dimensions=[257, 1024]

min: 233.35 s
max:  54.15 s
sum:  50.57 s
mean: 49.81 s