@ercius 2021-03-04
@psavery  Here are some answers:
> There is a dtype in the SparseArray class which indicates what
> dtype should be used for expansion and math functions. This
> currently defaults to np.uint8, which causes integer overflows
> for most of our examples, for things like summing. Should we
> increase the default to something else, like np.uint32?

Yes. I would definitely increase to uint32. uint8 is too small. Thanks!

> Adding a "bin_scans()" function might not be too hard. Should we
> add one of those too, to bin the scans?
Im not sure what bin_scans() would do. I thought the binning was already done in the array? A little more information is needed.